+++ scripts/modules/rust/cargo.sh
@@ -2,7 +2,7 @@
 
 # shellcheck source=/dev/null
 
-source /dev/stdin <<<"$(curl -s "https://raw.githubusercontent.com/dotbrains/utilities/master/scripts/base/base.sh")"
+smu::import base
 
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -