+++ scripts/modules/system/debian/apt.sh
@@ -3,7 +3,7 @@
 # shellcheck disable=SC2086
 # shellcheck source=/dev/null
 
-source /dev/stdin <<<"$(curl -s "https://raw.githubusercontent.com/dotbrains/utilities/master/scripts/base/base.sh")"
+smu::import base
 
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -