+++ scripts/modules/system/network.sh
@@ -2,7 +2,7 @@
 
 # shellcheck source=/dev/null
 
-source /dev/stdin <<<"$(curl -s "https://raw.githubusercontent.com/dotbrains/utilities/master/scripts/base/base.sh")"
+smu::import base
 
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
@@ -57,6 +57,7 @@ source_file() {
 
 	# - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -
 
-	source /dev/stdin <<<"$(curl -s "https://raw.githubusercontent.com/dotbrains/utilities/master/scripts/$TARGET")"
+	# Resolves local checkout, cache, then the pinned remote ref.
+	smu::load_file "$TARGET"
 
 }