+++ scripts/modules/system/system.sh
@@ -3,7 +3,7 @@
 # shellcheck source=/dev/null
 # shellcheck disable=2144,2010,2062,2063,2035
 
-source /dev/stdin <<<"$(curl -s "https://raw.githubusercontent.com/dotbrains/utilities/master/scripts/base/base.sh")"
+smu::import base
 
 # - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - - -